+++ controllers/keda/scaledobject_controller_test.go
@@ -719,7 +719,7 @@ var _ = Describe("ScaledObjectController", func() {
 		})
 	})
 
-	It("scaleobject ready condition 'False/Unknow' to 'True' will requeue", func() {
+	It("scaleobject ready condition 'False/Unknown' to 'True' will requeue", func() {
 		var (
 			deploymentName        = "conditionchange"
 			soName                = "so-" + deploymentName